Update PrivateDriverApp to use local time zone, and update dependencies
