+++ package.json
@@ -23,8 +23,10 @@
     "axios-retry": "^4.4.2",
     "expo": "^49.0.0",
     "expo-image": "~1.3.5",
+    "expo-localization": "~14.3.0",
     "expo-status-bar": "~1.6.0",
     "formik": "^2.2.9",
+    "moment-timezone": "^0.5.45",
     "react": "18.2.0",
     "react-native": "0.72.10",
     "react-native-calendars": "^1.1306.0",
@@ -36,6 +38,7 @@
     "react-native-text-link": "^1.0.0",
     "react-native-vector-icons": "^9.2.0",
     "react-navigation-tabs": "^2.11.2",
+    "rimraf": "^6.0.1",
     "styled-components": "^5.3.6",
     "typescript": "^4.6.3",
     "yup": "^1.2.0"
